Be more flexible about where we can find the split-point
Only looking in the extname doesn't work so well if we're not the last
processor in the chain. Plus, if we extract it from the '_splitN' part,
we don't need to define separate extension processors for each.

Also, default to the second part: the zeroth part is not very useful,
and you can just use the un-split file if you want the first.