Update splitter.rb
Fixed Issue of exceeding selectors.

The problem with the ranges is, that if the first rule that contributes to a range has more than 1 selector, the total amount of selectors in that range will exceed the limit.

For example: lets say we want to get the second split with the range [4096..8191]. In the current implementation, if there is a rule with 100 selectors that starts at selector index 4000, it will be part of the current split (because 4000 + 100 = 4100, which lies inside the range [4096..8191]). As a result the limit in this range could be exceeded by up to 95 selectors (especially if there are no selectors truncated at the end of the range). I therefor propose a rolling window.